+++ spec/airbrake_spec.rb
@@ -12,6 +12,17 @@ RSpec.describe Airbrake do
       end.to yield_with_args(Airbrake::Config)
     end
 
+    it "sets logger to Airbrake::Loggable" do
+      logger = Logger.new(File::NULL)
+      described_class.configure do |c|
+        c.project_id = 1
+        c.project_key = '123'
+        c.logger = logger
+      end
+
+      expect(Airbrake::Loggable.instance).to eql(logger)
+    end
+
     context "when user config doesn't contain a project id" do
       it "raises error" do
         expect { described_class.configure { |c| c.project_key = '1' } }.